requirements.txt
python-telegram-bot
ujson
unicodeit
